backend/requirements.txt
# Synapse 2.0 Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Dedalus Labs SDK (Agent Orchestration)
dedalus_labs[aiohttp]>=0.2.0

# OpenAI SDK (for K2 Think via OpenAI-compatible API)
openai>=1.0.0

# Snowflake
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1

# ElevenLabs
elevenlabs>=1.0.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Audio Processing
numpy==1.26.3
